+++ build.sh
@@ -2,4 +2,4 @@
 
 # Go to current directory so we can double click the file from anywhere
 cd "${0%/*}"
-cat ./src/barcode_header.js ./src/espm.js ./src/id_Util.js ./src/isbn.js ./src/barcode_library.js ./src/esfm.js ./src/barcode_ui.js ./src/barcode_drawer.js ./src/barcode_main.js | grep -Ev "#include" > id_barcode.jsx
+cat ./src/header.js ./src/presetManager.js ./src/idUtil.js ./src/isbn.js ./src/barcode_library.js ./src/fontDrop.js ./src/sui.js ./src/barcode_drawer.js ./src/main.js | grep -Ev "#include" > id_barcode.jsx
