+++ src/idUtil.js
